requirements.txt
asgiref==3.5.2
branca==0.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
folium==0.12.1.post1
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
